ai_learn/week_6/requirements.txt
numpy
pandas
scikit-learn
tqdm
matplotlib
sentence-transformers
faiss-cpu
python-dotenv
joblib